Separating ScalaZ and Cats pipeline
