+++ util-validators-cats/src/main/scala/com/outworkers/util/validators/cats/ValidatorImplicits.scala
@@ -1,10 +1,12 @@
-package com.outworkers.util.validators
+package com.outworkers.util.validators.cats
+
 
 import cats.data.Validated.{Invalid, Valid}
 import cats.data.{NonEmptyList, Validated, _}
 import cats.{Applicative, Semigroup, SemigroupK, _}
+import com.outworkers.util.validators.cats
 
-import scalaz.{Validation, _}
+import scalaz._
 
 trait ValidatorImplicits {
   implicit class ValidatedApiError[T](val valid: Validated[Map[String, List[String]], T]) {
@@ -42,34 +44,6 @@ trait ValidatorImplicits {
 
   implicit val nelSemigroup: Semigroup[NonEmptyList[(String, String)]] =
     SemigroupK[NonEmptyList].algebra[(String, String)]
-
-  implicit class ValidatedNelAugmenter[T](val v1: Nel[T]) {
-    def and[T2](v2: Nel[T2]): Wrapper2[T, T2] = Wrapper2[T, T2](v1, v2)
-  }
-
-  implicit class ScalazToCatsValidation[X, T](val vd: Validation[X, T]) {
-    def cats: Validated[X, T] = vd.fold(
-      fail => Invalid(fail),
-      valid => Valid(valid)
-    )
-  }
-
-  /**
-    * Augments cats validators with prop labelling.
-    * @param vd The validation to augment.
-    * @tparam T The underlying type of a successful validation.
-    */
-  implicit class CatsPropAugmenter[T](val vd: Validated[String, T]) {
-    def prop(str: String): ValidatedNel[(String, String), T] = {
-      vd.leftMap(f => str -> f).toValidatedNel
-    }
-  }
-
-  implicit class ScalazStringVdToCatsValidation[T](val vd: Validation[String, T]){
-    def prop(str: String): ValidatedNel[(String, String), T] = {
-      vd.leftMap(f => str -> f).cats.toValidatedNel
-    }
-  }
 
   implicit class ScalazToCatsValidationNel[X, T](val vd: ValidationNel[X, T]) {
