+++ util-validators-cats/src/main/scala/com/outworkers/util/validators/cats/Wrappers.scala
@@ -1,19 +1,5 @@
-/*
- * Copyright 2013 - 2017 Outworkers Ltd.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.outworkers.util.validators
+package com.outworkers.util.validators.cats
+
 
 import cats.Apply
 import cats.data.ValidatedNel
@@ -225,7 +211,7 @@ trait Wrappers {
     v11: Nel[T11],
     v12: Nel[T12]
   ) extends Wrapper[(T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12)] {
-    override def map[R](fn: ((T1, T2, T3, T4, T5, T6, T7, T8, T9)) => R): Nel[R] = {
+    override def map[R](fn: ((T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12)) => R): Nel[R] = {
       Apply[ValidatedNel[(String, String), ?]].map12[T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12,R](
         v1, v2, v3, v4, v5, v6, v7, v8, v9, v10, v11, v12
       ) { case x => fn(x) }
