+++ util-validators-cats/src/main/scala/com/outworkers/util/validators/cats/package.scala
@@ -1,0 +1,3 @@
+package com.outworkers.util.validators
+
+package object cats extends Wrappers with ValidatorImplicits
